Move toggleCardTable into case-table-card-utils.ts.
